Move various reductions to IndexedFrame
