+++ python/cudf/cudf/core/frame.py
@@ -31,7 +31,6 @@ from typing_extensions import Self
 import cudf
 from cudf import _lib as libcudf
 from cudf._typing import Dtype
-from cudf.api.extensions import no_default
 from cudf.api.types import is_bool_dtype, is_dtype_equal, is_scalar
 from cudf.core.buffer import acquire_spill_lock
 from cudf.core.column import (
@@ -1816,290 +1815,6 @@ class Frame(BinaryOperand, Scannable):
         )
 
     @_cudf_nvtx_annotate
-    def sum(
-        self,
-        axis=no_default,
-        skipna=True,
-        dtype=None,
-        numeric_only=False,
-        min_count=0,
-        **kwargs,
-    ):
-        """
-        Return sum of the values in the DataFrame.
-
-        Parameters
-        ----------
-        axis: {index (0), columns(1)}
-            Axis for the function to be applied on.
-        skipna: bool, default True
-            Exclude NA/null values when computing the result.
-        dtype: data type
-            Data type to cast the result to.
-        numeric_only : bool, default False
-            If True, includes only float, int, boolean columns.
-            If False, will raise error in-case there are
-            non-numeric columns.
-        min_count: int, default 0
-            The required number of valid values to perform the operation.
-            If fewer than min_count non-NA values are present the result
-            will be NA.
-
-            The default being 0. This means the sum of an all-NA or empty
-            Series is 0, and the product of an all-NA or empty Series is 1.
-
-        Returns
-        -------
-        Series
-
-        Examples
-        --------
-        >>> import cudf
-        >>> df = cudf.DataFrame({'a': [1, 2, 3, 4], 'b': [7, 8, 9, 10]})
-        >>> df.sum()
-        a    10
-        b    34
-        dtype: int64
-
-        .. pandas-compat::
-            **DataFrame.sum, Series.sum**
-
-            Parameters currently not supported are `level`, `numeric_only`.
-        """
-        return self._reduce(
-            "sum",
-            axis=axis,
-            skipna=skipna,
-            dtype=dtype,
-            numeric_only=numeric_only,
-            min_count=min_count,
-            **kwargs,
-        )
-
-    @_cudf_nvtx_annotate
-    def product(
-        self,
-        axis=no_default,
-        skipna=True,
-        dtype=None,
-        numeric_only=False,
-        min_count=0,
-        **kwargs,
-    ):
-        """
-        Return product of the values in the DataFrame.
-
-        Parameters
-        ----------
-        axis: {index (0), columns(1)}
-            Axis for the function to be applied on.
-        skipna: bool, default True
-            Exclude NA/null values when computing the result.
-        dtype: data type
-            Data type to cast the result to.
-        numeric_only : bool, default False
-            If True, includes only float, int, boolean columns.
-            If False, will raise error in-case there are
-            non-numeric columns.
-        min_count: int, default 0
-            The required number of valid values to perform the operation.
-            If fewer than min_count non-NA values are present the result
-            will be NA.
-
-            The default being 0. This means the sum of an all-NA or empty
-            Series is 0, and the product of an all-NA or empty Series is 1.
-
-        Returns
-        -------
-        Series
-
-        Examples
-        --------
-        >>> import cudf
-        >>> df = cudf.DataFrame({'a': [1, 2, 3, 4], 'b': [7, 8, 9, 10]})
-        >>> df.product()
-        a      24
-        b    5040
-        dtype: int64
-
-        .. pandas-compat::
-            **DataFrame.product, Series.product**
-
-            Parameters currently not supported are level`, `numeric_only`.
-        """
-
-        return self._reduce(
-            # cuDF columns use "product" as the op name, but cupy uses "prod"
-            # and we need cupy if axis == 1.
-            "prod" if axis in {1, "columns"} else "product",
-            axis=axis,
-            skipna=skipna,
-            dtype=dtype,
-            numeric_only=numeric_only,
-            min_count=min_count,
-            **kwargs,
-        )
-
-    # Alias for pandas compatibility.
-    prod = product
-
-    @_cudf_nvtx_annotate
-    def mean(self, axis=0, skipna=True, numeric_only=False, **kwargs):
-        """
-        Return the mean of the values for the requested axis.
-
-        Parameters
-        ----------
-        axis : {0 or 'index', 1 or 'columns'}
-            Axis for the function to be applied on.
-        skipna : bool, default True
-            Exclude NA/null values when computing the result.
-        numeric_only : bool, default False
-            If True, includes only float, int, boolean columns.
-            If False, will raise error in-case there are
-            non-numeric columns.
-        **kwargs
-            Additional keyword arguments to be passed to the function.
-
-        Returns
-        -------
-        mean : Series or DataFrame (if level specified)
-
-        Examples
-        --------
-        >>> import cudf
-        >>> df = cudf.DataFrame({'a': [1, 2, 3, 4], 'b': [7, 8, 9, 10]})
-        >>> df.mean()
-        a    2.5
-        b    8.5
-        dtype: float64
-        """
-        return self._reduce(
-            "mean",
-            axis=axis,
-            skipna=skipna,
-            numeric_only=numeric_only,
-            **kwargs,
-        )
-
-    @_cudf_nvtx_annotate
-    def std(
-        self,
-        axis=no_default,
-        skipna=True,
-        ddof=1,
-        numeric_only=False,
-        **kwargs,
-    ):
-        """
-        Return sample standard deviation of the DataFrame.
-
-        Normalized by N-1 by default. This can be changed using
-        the `ddof` argument
-
-        Parameters
-        ----------
-        axis: {index (0), columns(1)}
-            Axis for the function to be applied on.
-        skipna: bool, default True
-            Exclude NA/null values. If an entire row/column is NA, the result
-            will be NA.
-        ddof: int, default 1
-            Delta Degrees of Freedom. The divisor used in calculations
-            is N - ddof, where N represents the number of elements.
-        numeric_only : bool, default False
-            If True, includes only float, int, boolean columns.
-            If False, will raise error in-case there are
-            non-numeric columns.
-
-        Returns
-        -------
-        Series
-
-        Examples
-        --------
-        >>> import cudf
-        >>> df = cudf.DataFrame({'a': [1, 2, 3, 4], 'b': [7, 8, 9, 10]})
-        >>> df.std()
-        a    1.290994
-        b    1.290994
-        dtype: float64
-
-        .. pandas-compat::
-            **DataFrame.std, Series.std**
-
-            Parameters currently not supported are `level` and
-            `numeric_only`
-        """
-
-        return self._reduce(
-            "std",
-            axis=axis,
-            skipna=skipna,
-            ddof=ddof,
-            numeric_only=numeric_only,
-            **kwargs,
-        )
-
-    @_cudf_nvtx_annotate
-    def var(
-        self,
-        axis=no_default,
-        skipna=True,
-        ddof=1,
-        numeric_only=False,
-        **kwargs,
-    ):
-        """
-        Return unbiased variance of the DataFrame.
-
-        Normalized by N-1 by default. This can be changed using the
-        ddof argument.
-
-        Parameters
-        ----------
-        axis: {index (0), columns(1)}
-            Axis for the function to be applied on.
-        skipna: bool, default True
-            Exclude NA/null values. If an entire row/column is NA, the result
-            will be NA.
-        ddof: int, default 1
-            Delta Degrees of Freedom. The divisor used in calculations is
-            N - ddof, where N represents the number of elements.
-        numeric_only : bool, default False
-            If True, includes only float, int, boolean columns.
-            If False, will raise error in-case there are
-            non-numeric columns.
-
-        Returns
-        -------
-        scalar
-
-        Examples
-        --------
-        >>> import cudf
-        >>> df = cudf.DataFrame({'a': [1, 2, 3, 4], 'b': [7, 8, 9, 10]})
-        >>> df.var()
-        a    1.666667
-        b    1.666667
-        dtype: float64
-
-        .. pandas-compat::
-            **DataFrame.var, Series.var**
-
-            Parameters currently not supported are `level` and
-            `numeric_only`
-        """
-        return self._reduce(
-            "var",
-            axis=axis,
-            skipna=skipna,
-            ddof=ddof,
-            numeric_only=numeric_only,
-            **kwargs,
-        )
-
-    @_cudf_nvtx_annotate
     def all(self, axis=0, skipna=True, **kwargs):
         """
         Return whether all elements are True in DataFrame.
@@ -2214,61 +1929,6 @@ class Frame(BinaryOperand, Scannable):
             "any",
             axis=axis,
             skipna=skipna,
-            **kwargs,
-        )
-
-    def median(
-        self, axis=None, skipna=True, level=None, numeric_only=None, **kwargs
-    ):
-        """
-        Return the median of the values for the requested axis.
-
-        Parameters
-        ----------
-        axis : {index (0), columns (1)}
-            Axis for the function to be applied on. For Series this
-            parameter is unused and defaults to 0.
-        skipna : bool, default True
-            Exclude NA/null values when computing the result.
-        numeric_only : bool, default False
-            If True, includes only float, int, boolean columns.
-            If False, will raise error in-case there are
-            non-numeric columns.
-
-        Returns
-        -------
-        scalar
-
-        Examples
-        --------
-        >>> import cudf
-        >>> ser = cudf.Series([10, 25, 3, 25, 24, 6])
-        >>> ser
-        0    10
-        1    25
-        2     3
-        3    25
-        4    24
-        5     6
-        dtype: int64
-        >>> ser.median()
-        17.0
-
-        .. pandas-compat::
-            **DataFrame.median, Series.median**
-
-            Parameters currently not supported are `level` and `numeric_only`.
-
-        .. pandas-compat::
-            **DataFrame.median, Series.median**
-
-            Parameters currently not supported are `level` and `numeric_only`.
-        """
-        return self._reduce(
-            "median",
-            axis=axis,
-            skipna=skipna,
-            numeric_only=numeric_only,
             **kwargs,
         )
